finish all parts of assets screen, apis need updates to get full and proper data
