--- src/Components/Asset/index.ts
@@ -1,1 +1,0 @@
-export * from './Asset';
