+++ src/Components/index.ts
@@ -11,5 +11,5 @@ export * from './Sprite';
 export * from './SpriteSheet';
 export * from './Typography';
 export * from './FooterNav';
-export * from './Asset';
+export * from './AssetRow';
 export * from './Loading';
